+++ lib/elixir_lokalise_api/collections/permission_templates.ex
@@ -1,0 +1,4 @@
+defmodule ElixirLokaliseApi.Collection.PermissionTemplates do
+  @moduledoc false
+  defstruct items: []
+end