+++ lib/elixir_lokalise_api/models/permission_template.ex
@@ -1,0 +1,11 @@
+defmodule ElixirLokaliseApi.Model.PermissionTemplate do
+  @moduledoc false
+  defstruct id: nil,
+            role: nil,
+            permissions: [],
+            description: nil,
+            tag: nil,
+            tagColor: nil,
+            tagInfo: nil,
+            doesEnableAllReadOnlyLanguages: nil
+end